requirements.txt
aiohttp
configparser
future
numpy
openai
pipmaster
pydantic
python-dotenv
azure-storage-blob
httpx
azure-core
networkx
setuptools
tenacity
tiktoken
